circleci: fix default name should be build.
If you are using Workflows, jobs must have a name that is unique within the .circleci/config.yml file.

If you are not using workflows, the jobs map must contain a job named build.
This build job is the default entry-point for a run that is triggered by a push to your VCS provider.
It is possible to then specify additional jobs and run them using the CircleCI API